refactor(Loggging): 로깅 관련 설정을 common 모듈에서 관리하도록 변경
