+++ stempo-common/build.gradle.kts
@@ -3,6 +3,12 @@ dependencies {
     implementation(Dependencies.springBootStarterWeb)
     implementation(Dependencies.springBootStarterSecurity)
     implementation(Dependencies.springBootStarterDataJpa)
+    implementation(Dependencies.springAop)
+
+    // Logging
+    implementation(Dependencies.logbackCore)
+    implementation(Dependencies.logbackClassic)
+    implementation(Dependencies.logstashLogbackEncoder)
 
     // Util
     implementation(Dependencies.swagger)
