+++ stempo-common/src/main/java/com/stempo/config/AsyncConfig.java
@@ -1,5 +1,6 @@
 package com.stempo.config;
 
+import com.stempo.logging.decorator.MdcTaskDecorator;
 import java.util.concurrent.Executor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
