+++ stempo-common/src/main/java/com/stempo/logging/aspect/RepositoryLoggingAspect.java
@@ -1,6 +1,6 @@
-package com.stempo.aspect;
+package com.stempo.logging.aspect;
 
-import com.stempo.constants.MdcConstants;
+import com.stempo.logging.constants.MdcConstants;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
@@ -11,16 +11,16 @@ import org.springframework.stereotype.Component;
 @Aspect
 @Component
 @Slf4j
-public class ServiceLoggingAspect {
+public class RepositoryLoggingAspect {
 
-    @Around("execution(* com.stempo.service..*(..))")
-    public Object logServiceExecutionTime(ProceedingJoinPoint joinPoint) throws Throwable {
+    @Around("execution(* com.stempo.repository..*(..))")
+    public Object logRepositoryExecutionTime(ProceedingJoinPoint joinPoint) throws Throwable {
         long start = System.currentTimeMillis();
         try {
             return joinPoint.proceed();
         } finally {
             long duration = System.currentTimeMillis() - start;
-            MDC.put(MdcConstants.SERVICE_EXECUTION_TIME_MS, String.valueOf(duration));
+            MDC.put(MdcConstants.REPOSITORY_EXECUTION_TIME_MS, String.valueOf(duration));
         }
     }
 }
